♻️ Move supabase docker from nox to conftest
